refactor: shift hasProps to shared
