+++ src/app/modules/spcp/spcp.util.ts
@@ -3,6 +3,7 @@ import crypto from 'crypto'
 import { StatusCodes } from 'http-status-codes'
 import { err, ok, Result } from 'neverthrow'
 
+import { hasProp } from '../../../shared/util/has-prop'
 import {
   AuthType,
   BasicField,
@@ -11,7 +12,6 @@ import {
   SPCPFieldTitle,
 } from '../../../types'
 import { createLoggerWithLabel } from '../../config/logger'
-import { hasProp } from '../../utils/has-prop'
 import { MissingFeatureError } from '../core/core.errors'
 import {
   AuthTypeMismatchError,
