+++ src/shared/util/has-prop.ts
